feat: PotionBottle 障碍物 + LevelConfig byte[] 对齐 + 投射物障碍命中修复 + Spawner 优化
- PotionBottle (ObstacleType=11): 4色子瓶 bitmask 机制，逐色击破
- LevelConfig: GroundHealths/CoverHealths/ObstacleStages/ObstacleStates 统一为 byte[]（匹配 JSON base64 编码）
- SimulationMatchHandler: 移除 tile==None 早期跳过，投射物可命中无 tile 的障碍物格子
- RuleBasedSpawnModel: 移除 anti-streak（已被 Spawner 配置系统取代）+ 权重基数 100→1000
- 测试: anti-streak 测试移除 + base64 obstacleStages 解析测试 + byte[] 对齐

+++ src/Match3.Core.Tests/Analysis/EliminationPathConsistencyTests.cs
@@ -71,7 +71,7 @@ public class EliminationPathConsistencyTests
 
         // Grounds: 6 ice tiles in a 3x2 block at rows 3-4, cols 3-5
         var grounds = new GroundType[len];
-        var groundHealths = new int[len];
+        var groundHealths = new byte[len];
         for (int y = 3; y <= 4; y++)
         {
             for (int x = 3; x <= 5; x++)
